+++ shapiq/approximator/_base.py
@@ -413,7 +413,7 @@ class Approximator(ABC):
         )
 
 
-class ShapleyWeightsMixin(ABC):
+class ShapleySamplingMixin:
     """Mixin class for the computation of Shapley weights.
 
     Provides the common functionality for regression-based approximators like
@@ -425,7 +425,9 @@ class ShapleyWeightsMixin(ABC):
     def __init__(self) -> None:
         """Initializes the regression mixin."""
         self._big_M = float(1_000_000)
-        self.n: int
+        # check weather self is a subclass of Approximator
+        if not issubclass(self.__class__, Approximator):
+            raise TypeError("ShapleyWeightsMixin can only be used with subclasses of Approximator.")
 
     def _init_ksh_sampling_weights(self) -> np.ndarray[float]:
         """Initializes the weights for sampling subsets.
@@ -465,3 +467,57 @@ class ShapleyWeightsMixin(ABC):
         weights[np.logical_not(subsets).all(axis=1)] = self._big_M
         weights[subsets.all(axis=1)] = self._big_M
         return weights
+
+    def _sample_subsets(
+        self,
+        budget: int,
+        sampling_weights: np.ndarray[float],
+        replacement: bool = False,
+        pairing: bool = True,
+    ) -> np.ndarray[bool]:
+        """Samples subsets with the given budget.
+
+        Args:
+            budget: budget for the sampling.
+            sampling_weights: weights for sampling subsets of certain sizes and indexed by the size.
+                The shape is expected to be (n + 1,). A size that is not to be sampled has weight 0.
+            pairing: whether to use pairing (`True`) sampling or not (`False`). Defaults to `False`.
+
+        Returns:
+            sampled subsets.
+        """
+        # sanitize input parameters
+        sampling_weights = copy.copy(sampling_weights)
+        sampling_weights /= np.sum(sampling_weights)
+
+        # adjust budget for paired sampling
+        if pairing:
+            budget = budget - budget % 2  # must be even for pairing
+            budget = int(budget / 2)
+
+        # create storage array for given budget
+        subset_matrix = np.zeros(shape=(budget, self.n), dtype=bool)
+
+        # sample subsets
+        sampled_sizes = self._rng.choice(self.N_arr, size=budget, p=sampling_weights).astype(int)
+        if replacement:  # sample subsets with replacement
+            permutations = np.tile(np.arange(self.n), (budget, 1))
+            self._rng.permuted(permutations, axis=1, out=permutations)
+            for i, subset_size in enumerate(sampled_sizes):
+                subset = permutations[i, :subset_size]
+                subset_matrix[i, subset] = True
+        else:  # sample subsets without replacement
+            sampled_subsets, n_sampled = set(), 0  # init sampling variables
+            while n_sampled < budget:
+                subset_size = sampled_sizes[n_sampled]
+                subset = tuple(sorted(self._rng.choice(np.arange(0, self.n), size=subset_size)))
+                sampled_subsets.add(subset)
+                if len(sampled_subsets) != n_sampled:  # subset was not already sampled
+                    subset_matrix[n_sampled, subset] = True
+                    n_sampled += 1  # continue sampling
+
+        if pairing:
+            subset_matrix = np.repeat(subset_matrix, repeats=2, axis=0)  # extend the subset matrix
+            subset_matrix[1::2] = np.logical_not(subset_matrix[1::2])  # flip sign of paired subsets
+
+        return subset_matrix
